Add files via upload
The app.py file is the core component of the Streamlit web application that allows users to interact with the face mask detection model. It enables users to either upload an image or use a webcam to detect whether a person is wearing a face mask. The file integrates with OpenCV for real-time face detection and MediaPipe for facial landmark extraction. Once an image is uploaded or captured, the app processes the image, applies the trained CNN model, and displays the prediction results. The app provides a simple and user-friendly interface for real-time mask detection, making it accessible and intuitive for users without requiring any technical expertise.